Use isSubtypeOf rather than typesSatisfyConstraint.
The former is for subtype checks. The latter can be used for a variety
of constraint checks, and has extra flexibility, e.g. the ability to
open archetypes. Opening archetypes for subtype checks does not make
sense, though, since we end up with constraint systems that cannot be
solved because they potentially lack the type context that would be
required in order to choose bindings that resolve all of the
constraints.

Fixes rdar://problem/32651746 / SR-5164.