[Type checker] Always record used conformances in the source file.
An early exit introduced in PR #10292 meant that we would only record
a protocol conformance as "used" in the first file in which we saw a
reference to the conformance. Make sure we record the conformance as
used in each source file that needs it. Fixes rdar://problem/32978891
and rdar://problem/32980176.