SILGen: Fix linkage for stored property initializers of private properties
Always give stored property initializers the linkage of the type
that contains them, even if the property is less visible than the
type.

This fixes the case where a type defines a private property, and an
extension in a different file from the same module defines a
constructor.

Fixes <rdar://problem/32743703>.