[QoI] Improve the fix-it for out-of-order arguments
Change the fix-it to move the argument to its correct location in one go.
This happens by removing it from one location and inserting it in the other
(as opposed to the original implementation which swapped one argument with
the preceding one). The commas separating the arguments are adjusted
to match the moved argument.

Add new tests for reordering regular arguments, variadic arguments, and function arguments.

Resolves: SR-4715 rdar://problem/31849281
(cherry picked from commit 00833a234309103dba19d487567cf6fe5d510528)