[sil-ownership] Allow for switch_enum to take a guaranteed parameter.
This is needed to ensure that SILGenPattern propagates values at +0 instead of
+1.

I also improved our handling of terminators in general when it comes to having
"borrowed" arguments. The basic verification requirement is that the
end_borrow_argument is treated as a use of the parent borrow. This ensures that
the borrowed argument can not extend the lifetime of the borrowed value without
the verifier triggering.

rdar://31880847