[ClangImporter] Structs lexically in an ObjC class are still top-level. (#10874)
Not every declaration that's syntactically in an Objective-C container
is a member of that container. Double-check the decl context before
adding it.

This is technically a breaking change in non-asserts builds, because
the struct really could be found via member lookup. But that should
be considered a bug, and I /suspect/ no one is relying on it.

rdar://problem/32451417