[Constraint system] Improve SE-0110 exception by allowing optional injection.
There was an oversight in the exception that was added to SE-0110 to
maintain compatibility for closure arguments. We were not allowing
optional injection to happen for the closure being passed in, so things
like:
  func test(_ fn: ((Int, Int))->()) {}
  test { x, y in }
worked, but
  func test(_ fn: (((Int, Int))->())?) {}
  test { x, y in }
did not.

Fixes SR-5433 (rdar://problem/33239714).

(cherry picked from commit 8605b29951cf27c36c1bb26b548efe1c754e9927)