Sema: Explicitly infer requirements from subscript element type for materializeForSet
The interface type for materializeForSet does not mention the
element type of a subscript, so Sema was not inferring
requirements that were implied by the element type but not
explicitly stated in the generic signature.

This led to assertion failures in SILGen because the
materializeForSet calls the getter and setter, which does
have these requirements.

Fixes <rdar://problem/33219034>, <https://bugs.swift.org/browse/SR-5420>.