Sema: Add test case for lazy property 'resetting' that no longer works
Due to a serendipitous confluence of interesting behaviors,
Swift 3 allowed you to define a lazy property with an IUO
type, and setting the property to nil would 'reset' it,
re-evaluating the lazy getter the next time the property
was accessed.

This was not intended behavior and it no longer works,
so add a test that it no longer works.

Fixes <rdar://problem/32687168> and
<https://bugs.swift.org/browse/SR-5172>.