[stdlib] Actually build in Swift 4 mode
This is not ready for merge yet because, at the very least, we don't (yet) have
an answer for allZeros that doesn't force us to keep it around forever.  As long
as we build the standard library in Swift 3.2 mode, the compiler's current
combination of bugs and features allows us to prevent users from actually using
it when they build in Swift 4 mode.