[Driver] Hack: Add -Xcc -iquote to a bridging header's directory.
It turns out that Clang's been happy to pick up module maps from the
same directory as a bridging header imported with -import-objc-header.
People were implicitly relying on this behavior, which broke when we
started precompiling bridging headers---suddenly the "same directory"
was wherever the precompiled form was (either a temporary directory or
a persistent build intermediate). Emulate the old behavior by formally
making this directory a Clang search path using the -iquote option.
Why -iquote? Because (a) it always goes first, and (b) it won't disturb
textual includes that use angle-brackets, while still getting Clang to
check the directory for a module map.

We may want to change this behavior in the future, but for now it's
more important to keep things consistent between precompiled and
non-precompiled bridging headers, and the easiest way to do that is to
follow the old behavior.

rdar://problem/32810754