Disable static knowledge of graphemes.
With the standard library now relying on ICU for grapheme breaking,
what is and is not a valid grapheme is unknowable at compilation time
as the supported Unicode version is supplied at run time. Rather than
give incorrect errors, or even downgrade to un-suppressable warnings,
all non-empty Strings are considered to potentially be graphemes.

Future directions could include warnings/errors for known stable
non-graphemes, multi-version grapheme breaking implementations coupled
with deployment-target checking, etc. In any case, we should be
careful to not give users the false impression that static checks
are sufficient/exhaustive.

This commit is meant to be minimally invasive. A subsequent commit
will more aggressively strip the compiler of grapheme knowledge.