[stdlib] Strip allocator implementation dependencies from a test
These tests depend on the specific characteristics of the memory allocator under
use.  In particular, if the allocator decides to return a larger memory block
than requested, the tests may fail.

I can only make wild guesses as to why changes to String would cause this test
to fail on Linux, but that's what we're observing.

  check failed at .../swift/validation-test/stdlib/Inputs/CommonArrayTests.gyb, line 159
  8 >= 9