[sil-opened-archetypes-tracker] Remove only those archetype definitions that are registered with the current SILOpenedArchetypesTracker
It may happen that a valid opened archetype was not registered in the SILOpenedArchetypesTracker, because the tracker was created without scanning the whole function and thus may not aware of all opened archetypes of the function. This often happens e.g. during inlining. Therefore, remove the archetype definition only if it was registered before.

Fixes rdar://problem/33513110